requirements.txt
## use python=3.10
streamlit
google-generativeai
python-dotenv
langchain==0.0.352
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
